Move editor UI functions to `editor_ui.cpp` and reorder
